Halton Iterative Partitioning (HIP).

HIP draws spatially balanced samples and over-samples from point resources by partitioning the resource into boxes with the same nested structure as Halton boxes. The spbal parameter iterations defines the number of boxes used in the HIP partition and should be larger than the sample size but less than the population size. The iterations parameter also defines the number of units available in the HIP over-sample, where the over-sample contains one unit from each box in the HIP partition.

Usage

HIP(
  population = NULL,
  n = 20,
  iterations = 7,
  minRadius = NULL,
  panels = NULL,
  panel_overlap = NULL,
  verbose = FALSE
)

Arguments

population

A population of point pairs.

n

The number of points to draw from the population. Default 20.

iterations

The levels of partitioning required. Default 7.

minRadius

If specified, the minimum distance, in meters, allowed between sample points. This is applied to the $overSample.

panels

A list of integers that define the size of each panel in a non-overlapping panels design. The length of the list determines the number of panels required. The sum of the integers in the panels parameter will determine the total number of samples selected, n. The default value for panels is NULL, this indicates that a non-overlapping panel design is not wanted.

panel_overlap

A list of integers that define the overlap into the previous panel. Is only used when the panels parameter is not NULL. The default value for panel_overlap is NULL. The length of panel_overlap must be equal to the length of panels. The first value is always forced to zero as the first panel never overlaps any region.

verbose

Boolean if you want to see any output printed to screen. Helpful if taking a long time. Default is FALSE i.e. no informational messages are displayed.

Value

Return a list containing the following five variables:

Population Original population point pairs as an sf object.

HaltonIndex The Halton index for the point. Points will be spread equally across all Halton indices.

sample The sample from the population of size n.

overSample The overSample contains one point from each Halton box. All contiguous sub-samples from oversample are spatially balanced, and the first n points are identical to sample.

minRadius This result variable contains the sample created using the minRadius parameter. If the minRadius parameter is not specified then the minRadius variable will contain NULL.

Details

Halton iterative partitioning (HIP) extends Basic acceptance sampling (BAS) to point resources. It partitions the resource into $B >= n$ boxes that have the same nested structure as in BAS, but different sizes. These boxes are then uniquely numbered using a random-start Halton sequence of length $B$. The HIP sample is obtained by randomly drawing one point from each of the boxes numbered $1, 2, . . . , n$.

Author

Phil Davies, Blair Robertson.

Examples

# generating 20 points from a population of 5,000 (random) points with 7
# levels of partitioning (4 in the first dimension and 3 in the second) to
# give (2^4) * (3^3) = 32 * 27, resulting in 864 boxes ----------------------

# set random seed
set.seed(511)

# define HIP parameters.
pop <- matrix(runif(5000*2), nrow = 5000, ncol = 2)
n <- 20
its <- 7

# Convert the population matrix to an sf point object.
sf_points <- sf::st_as_sf(data.frame(pop), coords = c("X1", "X2"))
dim(sf::st_coordinates(sf_points))
#> [1] 5000    2

# generate HIP sample.
result <- spbal::HIP(population = sf_points,
                     n = n,
                     iterations =  its)

# HaltonIndex
HaltonIndex <- result$HaltonIndex
table(HaltonIndex)
#> HaltonIndex
#>   0   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#>  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#>  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#>  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#>  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 120 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 140 141 142 143 144 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 160 161 162 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 180 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198 199 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 200 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218 219 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 220 221 222 223 224 225 226 227 228 229 230 231 232 233 234 235 236 237 238 239 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 240 241 242 243 244 245 246 247 248 249 250 251 252 253 254 255 256 257 258 259 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 260 261 262 263 264 265 266 267 268 269 270 271 272 273 274 275 276 277 278 279 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 280 281 282 283 284 285 286 287 288 289 290 291 292 293 294 295 296 297 298 299 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 300 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318 319 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 320 321 322 323 324 325 326 327 328 329 330 331 332 333 334 335 336 337 338 339 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 340 341 342 343 344 345 346 347 348 349 350 351 352 353 354 355 356 357 358 359 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 360 361 362 363 364 365 366 367 368 369 370 371 372 373 374 375 376 377 378 379 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 380 381 382 383 384 385 386 387 388 389 390 391 392 393 394 395 396 397 398 399 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 400 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418 419 
#>  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11  11 
#> 420 421 422 423 424 425 426 427 428 429 430 431 
#>  11  11  11  11  11  11  11  11  11  11  11  11 

# Population Sample
HIPsample <- result$sample
HIPsample
#> Simple feature collection with 20 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 0.04282557 ymin: 0.05113006 xmax: 0.9548381 ymax: 0.9662592
#> CRS:           NA
#> First 10 features:
#>    HaltonIndex spbalSeqID                     geometry
#> 1            0          1  POINT (0.9104533 0.1395789)
#> 2            1          2  POINT (0.2864811 0.6609214)
#> 3            2          3   POINT (0.590989 0.9033066)
#> 4            3          4  POINT (0.1926398 0.2614476)
#> 5            4          5  POINT (0.7768502 0.3905133)
#> 6            5          6  POINT (0.4038541 0.8324469)
#> 7            6          7 POINT (0.7109394 0.05113006)
#> 8            7          8 POINT (0.08206315 0.4703658)
#> 9            8          9  POINT (0.9548381 0.6773421)
#> 10           9         10  POINT (0.3318416 0.1603996)