Create a Halton Frame.

Halton frames discretize an areal resource into a spatially ordered grid, where samples of consecutive frame points are spatially balanced. To generate Halton Frames, spbal requires a study region shapefile and the region’s bounding box.

Usage

HaltonFrame(
  N = 1,
  J = base::c(3, 2),
  bases = base::c(2, 3),
  boundingbox = NULL,
  shapefile = NULL,
  panels = NULL,
  panel_overlap = NULL,
  seeds = NULL,
  stratum = NULL,
  verbose = FALSE
)

Arguments

N

The number of points in the frame to generate.

J

The number of grid cells. A list of 2 values. The default value is c(3, 2).

bases

Co-prime base for the Halton Sequence. The default value is c(2, 3).

boundingbox

Bounding box around the study area. If a bounding box is not supplied then spbal will generate a bounding box for the shapefile.

shapefile

A sf object. If the shapefile parameter is NULL then function HaltonFrameBase is called directly.

panels

A list of integers that define the size of each panel in a non-overlapping panels design. The length of the list determines the number of panels required. The sum of the integers in the panels parameter will determine the total number of samples selected, n. The default value for panels is NULL, this indicates that a non-overlapping panel design is not wanted.

panel_overlap

A list of integers that define the overlap into the previous panel. Is only used when the panels parameter is not NULL. The default value for panel_overlap is NULL. The length of panel_overlap must be equal to the length of panels. The first value is always forced to zero as the first panel never overlaps any region.

seeds

A vector of 2 seeds, u1 and u2. If not specified, the default is NULL.

stratum

Name of column in shapefile that makes up the strata.

verbose

Boolean if you want to see any output printed to screen. Helpful if taking a long time. Default is FALSE i.e. no informational messages are displayed.

Value

Returns a list containing five variables:

J The number of grid cells. A list of 2 values that were used to generate this Halton grid and frame.

hg.pts.shp Halton grid over the bounding box and study area.

hf.pts.shp Halton frame, the sample points within the study area.

bb The bounding box.

seeds The u1 and u2 seeds used to generate the sample.

The sample points in hf.pts.shp are returned in the form of a simple feature collection of POINT objects. As well as having the features from the original shapefile, the following new attributes have been added:

spbalSeqID: A unique identifier for every sample point.

ID: A unique identifier, the Halton frame point order.

Author

Phil Davies.

Examples

# we discretize the Gates study region into a coarse grid using
# B = 2^{J_1} * 3^{J_2}= (2^3) * (3^2) (9 by 8 grid) ------------------------

# Use the North Carolina shapefile supplied in the sf R package.
shp_file <- sf::st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/home/phil/R/x86_64-pc-linux-gnu-library/4.3/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
shp_gates <- shp_file[shp_file$NAME == "Gates",]

# Vertically aligned master sample bounding box.
bb <- spbal::BoundingBox(shapefile = shp_gates)

set.seed(511)
result6 <- spbal::HaltonFrame(shapefile = shp_gates,
                              J = c(3, 2),
                              boundingbox = bb)
# get the frame points.
Frame <- result6$hf.pts.shp
Frame
#> Simple feature collection with 52 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -76.92284 ymin: 36.309 xmax: -76.49118 ymax: 36.54077
#> Geodetic CRS:  NAD27
#> First 10 features:
#>    ID spbalSeqID                          x
#> 1   1          1  POINT (-76.67618 36.5118)
#> 2   2          2 POINT (-76.79951 36.36694)
#> 3   3          3 POINT (-76.55285 36.45386)
#> 4   4          4 POINT (-76.86118 36.54077)
#> 6   6          5 POINT (-76.73784 36.39591)
#> 9   9          6 POINT (-76.67618 36.42488)
#> 10 10          7  POINT (-76.79951 36.5118)
#> 11 11          8 POINT (-76.55285 36.36694)
#> 12 12          9 POINT (-76.86118 36.45386)
#> 13 13         10 POINT (-76.61451 36.54077)
# get the grid points.
Grid <- result6$hg.pts.shp
Grid
#> Simple feature collection with 72 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -76.92284 ymin: 36.309 xmax: -76.49118 ymax: 36.54077
#> Geodetic CRS:  NAD27
#> First 10 features:
#>                             x ID
#> 1   POINT (-76.67618 36.5118)  1
#> 2  POINT (-76.79951 36.36694)  2
#> 3  POINT (-76.55285 36.45386)  3
#> 4  POINT (-76.86118 36.54077)  4
#> 5    POINT (-76.61451 36.309)  5
#> 6  POINT (-76.73784 36.39591)  6
#> 7  POINT (-76.49118 36.48283)  7
#> 8  POINT (-76.92284 36.33797)  8
#> 9  POINT (-76.67618 36.42488)  9
#> 10  POINT (-76.79951 36.5118) 10