Extract all points with a specified panel id from a sample.

This is the main function for selecting sites using the BAS master sample. It assumes that you have already defined the master sample using the BoundingBox() function or will be selecting a marine master sample site in BC.

Usage

getPanel(shapefile, panelid)

Arguments

shapefile

Shape file as a polygon (sp or sf) containing a sample that contains a feature column named panel_id.

panelid

The overlapped panel in the shapefile shp the user wants sample points from.

Value

The sample for the specified panel.

Author

Phil Davies.

Examples

# Halton frame overlapping panel design showing use of getPanel.

# Use the North Carolina shapefile supplied in the sf R package.
shp_file <- sf::st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/home/phil/R/x86_64-pc-linux-gnu-library/4.3/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
shp_gates <- shp_file[shp_file$NAME == "Gates",]

# Vertically aligned master sample bounding box.
bb <- spbal::BoundingBox(shapefile = shp_gates)

# Three panels, of 20 samples each.
panels <- c(20, 20, 20)

# second panel overlaps first panel by 5, and third panel
# overlaps second panel by 5.
panel_overlap <- c(0, 5, 5)

# generate the sample.
samp <- spbal::HaltonFrame(J = c(4, 3),
                           boundingbox = bb,
                           panels = panels,
                           panel_overlap = panel_overlap,
                           shapefile = shp_gates)

# get halton frame data from our sample.
samp3 <- samp$hf.pts.shp
samp3
#> Simple feature collection with 309 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -76.93826 ymin: 36.29935 xmax: -76.47577 ymax: 36.55042
#> Geodetic CRS:  NAD27
#> First 10 features:
#>    ID spbalSeqID                          x panel_id
#> 1   1          1 POINT (-76.66076 36.45386)        1
#> 2   2          2 POINT (-76.78409 36.54077)        1
#> 4   4          3 POINT (-76.84576 36.40557)        1
#> 5   5          4  POINT (-76.5991 36.49248)        1
#> 6   6          5 POINT (-76.72243 36.34763)        1
#> 9   9          6   POINT (-76.6916 36.3766)        1
#> 10 10          7 POINT (-76.81493 36.46351)        1
#> 11 11          8 POINT (-76.56826 36.55042)        1
#> 13 13          9 POINT (-76.62993 36.38626)        1
#> 14 14         10 POINT (-76.75326 36.47317)        1

panelid <- 1
olPanel_1 <- spbal::getPanel(samp3, panelid)